1.02.06 - Reset
Setup the Reset action that will take the players back to the Splash screen by either one of the players pressing and holding the Select button for 5 seconds or if they leave the game paused for 5 minutes w/o pressing a button during a match.

Also send the players back to the Splash screen on the Ship Select Menu if the game is left untouched for a minute.

Modify the ChangeSceneOnButton script so it works on the Splash, Main Menu, and Control screens. The script move to the next scene if any button is pressed or after a specified amount of time. Have also given the option to return to the Splash screen after a specified amount of time. This feature was mainly made for the Main Menu screen and will return to the Splash screen after the main song finishes.